+++ src/types/request/transfers.ts
@@ -21,6 +21,7 @@ export type DepositRequest = {
 };
 export type CreateDepositProps = DepositRequest & {
   player_id: string;
+  coin_transfer_id: string;
   cep_ok?: boolean;
 };